// apps/web/src/lib/api/generated/types/createCandidateContactHistoryDtoOutreachChannel.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Linktal Recruitment API
 * API for Linktal Recruitment OS
 * OpenAPI spec version: 1.0
 */

/**
 * Which channel this outreach used — set when category is OUTREACH
 */
export type CreateCandidateContactHistoryDtoOutreachChannel = typeof CreateCandidateContactHistoryDtoOutreachChannel[keyof typeof CreateCandidateContactHistoryDtoOutreachChannel];


export const CreateCandidateContactHistoryDtoOutreachChannel = {
  SMS: 'SMS',
  EMAIL: 'EMAIL',
} as const;